Fix bugs in loading and saving to files
