+++ data/jerry.txt
@@ -1,1 +1,3 @@
-tasklist.TaskList@63c12fb0
+E | 0 | event play lalaa | 2020-10-10 12:30 | 2020-10-11 12:31
+T | 1 | lol
+D | 1 | deadline quiz /by 2022-10-10 12:30 | 2022-10-10 12:30
