+++ src/main/java/command/ByeCommand.java
@@ -1,7 +1,10 @@
+package command;
+
 import exceptions.JerryException;
 
 import storage.Storage;
 import tasklist.TaskList;
+import ui.Ui;
 
 
 public class ByeCommand extends Command {
@@ -11,7 +14,8 @@ public class ByeCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasklist, Ui ui, Storage storage) throws JerryException {
+    public void execute(TaskList taskList, Ui ui, Storage storage) throws JerryException {
+        taskList.saveTasks(storage);
         ui.displayOutput(this.end());
     }
