+++ src/main/java/command/DeleteCommand.java
@@ -1,13 +1,15 @@
+package command;
+
 import exceptions.InvalidCommandFormatException;
 import exceptions.JerryException;
 import storage.Storage;
 import tasklist.TaskList;
+import ui.Ui;
 
-public class MarkCommand extends Command {
-
+public class DeleteCommand extends Command {
     private int index = 0;
 
-    public MarkCommand(String input) throws InvalidCommandFormatException {
+    public DeleteCommand(String input) throws InvalidCommandFormatException {
         String[] entries = input.split(" ", 2);
         if (entries.length < 2 || entries[1].trim().isEmpty()) {
             throw new InvalidCommandFormatException("Task number must be positive!");
@@ -21,8 +23,8 @@ public class MarkCommand extends Command {
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws JerryException {
-        this.response = taskList.mark(index);
-        storage.writeToFile(taskList.toString());
+        this.response = taskList.deleteTask(index);
+        taskList.saveTasks(storage);
         ui.displayOutput(this.response);
     }
